Fixes wrong solution in Missing Values Checkpoint
Note I fixed it, showed some github issues of the problem happening.

I did *not* make it the most beautiful code ever (there are ways to make it look way better) nor did I include a huge warning against the use of inplace in Pandas, which is frowned upon. There are 3 issues in Pandas Dev (that I could find) in which that problem resurfaces. It also appeared in TC and it took some time for us to solve the issue, as it feels weird at first.